move Section & Header to Container
