+++ src/features/MoviesPage/index.js
@@ -3,12 +3,11 @@ import { useSelector, useDispatch } from "react-redux";
 import { MediumTile } from "../../common/Tile/MediumTile";
 import Pagination from "./../../common/Pagination/index";
 import { StatusChecker } from "./../../common/StatusChecker/index";
-import { Container } from "./../../common/Container/index";
+import { Container, Header, Section } from "./../../common/Container/index";
 import { NoResult } from "./../../common/NoResult";
 import { pageState } from "./../../common/pageState";
 import { selectGenerateList } from "./../../common/commonSlice";
 import { usePageParameter } from "./../usePageParameters";
-import { Header, Section } from "./styled";
 import { apiImage, apiKey } from "../../common/commonValues";
 import noPoster from "../../assets/noPoster.svg";
 import { popularMovies } from "./popularMovies";
