--- src/features/MoviesPage/styled.js
@@ -1,23 +1,0 @@
-import styled from "styled-components";
-
-export const Section = styled.section`
-display: flex;
-flex-wrap: wrap;
-gap: 12px;
-width: 100%;
-max-width: 1368px;
-justify-content: center;
-`;
-
-export const Header = styled.h2`
-  margin: 56px 0 24px;
-  font-weight: 600;
-  font-size: 36px;
-  line-height: 1.2;
-  color: ${({ theme }) => theme.colors.woodsmoke};
-
-  @media (max-width: ${({ theme }) => theme.breakpoints.tablet}px) {
-    font-size: 18px;
-    margin: 8px 0 12px 0;
-  }
-`;
